Added simple deploy script to prepare architecture installation directory
The script compresses the .so files of each node and stores them in
`install/<arch>/NodeName.tar.gz`.
The script is currently hardcoded to `x64` only, but we'll extend it as soon as x-compile
is set up and running.

The `deploy_arch.sh` script is called from the `build.sh` and invoked in the context of the
host machine.